WEB-2039 extract actions sheet
